requirements.txt
arrow==1.2.3
asgiref==3.6.0
binaryornot==0.4.4
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.4
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.6
django-browser-reload==1.6.0
django-cors-headers==3.14.0
django-environ==0.9.0
django-filter==23.1
django-tailwind @ git+https://github.com/timonweb/django-tailwind.git@37b81f631edddac04844413bb5635a1c73e5af30
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyTelegramBotAPI==3.7.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.4.0
